pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fpdf2>=2.8.5",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pandas-datareader>=0.10.0",
    "plotly>=6.3.1",
    "prophet>=1.2.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "statsmodels>=0.14.5",
    "streamlit>=1.51.0",
    "yfinance>=0.2.66",
]
